# PERF: MultiIndex.size