--- a/pandas/core/indexes/multi.py
+++ b/pandas/core/indexes/multi.py
@@ -786,6 +786,14 @@ class MultiIndex(Index):
     def __len__(self) -> int:
         return len(self.codes[0])
 
+    @property
+    def size(self) -> int:
+        """
+        Return the number of elements in the underlying data.
+        """
+        # override Index.size to avoid materializing _values
+        return len(self)
+
     # --------------------------------------------------------------------
     # Levels Methods